SmartCS_Desktop/server/ocr_engine.py
import base64
import io
from PIL import Image

# In a real production environment, you would import paddleocr or pytesseract here.
# from paddleocr import PaddleOCR
# ocr = PaddleOCR(use_angle_cls=True, lang='ch')

class OCREngine:
    def __init__(self):
        print("👁️ Visual Sentry Engine Initialized")

    def analyze_image(self, image_bytes):
        """
        Mock OCR implementation for demonstration.
        In reality, this returns text found in the image.
        """
        # 1. Convert bytes to image to ensure it's valid
        try:
            image = Image.open(io.BytesIO(image_bytes))
            # image.save("debug_ocr.png") # Debugging
        except:
            return ""

        # 2. Real OCR Logic would go here.
        # text = pytesseract.image_to_string(image, lang='chi_sim')
        
        # --- SIMULATION FOR DEMO ---
        # Since we cannot easily install heavy OCR libs in this environment without system deps,
        # we will simulate detection based on a "magic trigger" or just return empty
        # unless we implement a very basic pixel check or rely on Agent's self-test.
        
        # However, to make the user's "Test" work, let's assume the client 
        # sends a specific "Test Image" or we just return a simulation result 
        # if the file size matches a certain criteria or just random for demo?
        
        # Better: We just return "Nothing" by default, but if the user uploads 
        # a specific test image or we can't really do OCR without Tesseract installed on system.
        
        # Let's return a placeholder text that forces the client to handle logic.
        return "" 
    
    def mock_detect(self, keywords):
        """
        Helper for demo: If we were really running OCR, we'd return the text.
        """
        pass

ocr_engine = OCREngine()
