SmartCS_Desktop/requirements.txt
flet>=0.21.0
pynput>=1.7.6
sqlalchemy>=2.0.25
websockets>=12.0
requests>=2.31.0
cryptography>=42.0.0
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.9
Pillow>=10.2.0
python-dotenv>=1.0.0
redis>=5.0.0
pymysql>=1.1.0